feat(front): remove headband
